+++ src/bitstamp/mod.rs
@@ -1,2 +1,2 @@
 pub mod api;
-pub mod tools;
+pub mod utils;
