+++ src/bitstamp/utils.rs
